+++ src/client/sync.rs
@@ -149,12 +149,12 @@ impl Client {
     /// # Example
     /// ```no_run
     /// use ibapi::Client;
-    /// use ibapi::contracts::Contract;
+    /// use ibapi::contracts::{Contract, SecurityType};
     ///
     /// let client = Client::connect("127.0.0.1:4002", 100).expect("connection failed");
     /// let mut contract = Contract::default();
     /// contract.symbol = "AAPL".to_string();
-    /// contract.sec_type = "STK".to_string();
+    /// contract.security_type = SecurityType::Stock;
     /// contract.exchange = "SMART".to_string();
     /// contract.currency = "USD".to_string();
     ///
