+++ src/orders/builder/sync_impl.rs
@@ -4,6 +4,9 @@ use crate::contracts::Contract;
 use crate::errors::Error;
 use crate::orders;
 
+#[cfg(test)]
+mod tests;
+
 impl<'a> OrderBuilder<'a, Client> {
     /// Submit the order synchronously
     /// Returns the order ID assigned to the submitted order
@@ -88,19 +91,19 @@ impl Client {
     /// # Example
     /// ```no_run
     /// use ibapi::Client;
-    /// use ibapi::contracts::Contract;
+    /// use ibapi::contracts::{Contract, SecurityType};
     ///
     /// let client = Client::connect("127.0.0.1:4002", 100).expect("connection failed");
     ///
     /// let mut contract1 = Contract::default();
     /// contract1.symbol = "AAPL".to_string();
-    /// contract1.sec_type = "STK".to_string();
+    /// contract1.security_type = SecurityType::Stock;
     /// contract1.exchange = "SMART".to_string();
     /// contract1.currency = "USD".to_string();
     ///
     /// let mut contract2 = Contract::default();
     /// contract2.symbol = "MSFT".to_string();
-    /// contract2.sec_type = "STK".to_string();
+    /// contract2.security_type = SecurityType::Stock;
     /// contract2.exchange = "SMART".to_string();
     /// contract2.currency = "USD".to_string();
     ///
