+++ src/orders/builder/validation.rs
@@ -62,82 +62,4 @@ pub fn validate_stop_price(action: &Action, stop_price: f64, current_price: Opti
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_valid_buy_bracket() {
-        let result = validate_bracket_prices(
-            Some(&Action::Buy),
-            50.0, // entry
-            55.0, // take profit (above entry)
-            45.0, // stop loss (below entry)
-        );
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_valid_sell_bracket() {
-        let result = validate_bracket_prices(
-            Some(&Action::Sell),
-            50.0, // entry
-            45.0, // take profit (below entry)
-            55.0, // stop loss (above entry)
-        );
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_invalid_buy_bracket_take_profit() {
-        let result = validate_bracket_prices(
-            Some(&Action::Buy),
-            50.0, // entry
-            45.0, // take profit (BELOW entry - invalid)
-            45.0, // stop loss
-        );
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_invalid_buy_bracket_stop_loss() {
-        let result = validate_bracket_prices(
-            Some(&Action::Buy),
-            50.0, // entry
-            55.0, // take profit
-            55.0, // stop loss (ABOVE entry - invalid)
-        );
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_invalid_sell_bracket() {
-        let result = validate_bracket_prices(
-            Some(&Action::Sell),
-            50.0, // entry
-            55.0, // take profit (ABOVE entry - invalid for sell)
-            45.0, // stop loss (BELOW entry - invalid for sell)
-        );
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_stop_price_validation_buy() {
-        // Buy stop must be above current price
-        assert!(validate_stop_price(&Action::Buy, 55.0, Some(50.0)).is_ok());
-        assert!(validate_stop_price(&Action::Buy, 45.0, Some(50.0)).is_err());
-    }
-
-    #[test]
-    fn test_stop_price_validation_sell() {
-        // Sell stop must be below current price
-        assert!(validate_stop_price(&Action::Sell, 45.0, Some(50.0)).is_ok());
-        assert!(validate_stop_price(&Action::Sell, 55.0, Some(50.0)).is_err());
-    }
-
-    #[test]
-    fn test_stop_price_no_current_price() {
-        // Should pass if no current price provided
-        assert!(validate_stop_price(&Action::Buy, 55.0, None).is_ok());
-        assert!(validate_stop_price(&Action::Sell, 45.0, None).is_ok());
-    }
-}
+mod tests;
